200 transducerst hit 30x100 um spot.
From Riley's email:

Transducer array of 50 micron sized transducers (with 50 microns
of spacing between them), I simulated 200 transducers together along a
curved surface which approximates the inside of the skull (6.5 cm radius
of curvature). This is with 40 MHz.

This image shows focusing on a spot 1cm away from the array, and the
bright little dot we're hitting is 30x100 microns. That's the size of
single neuron! Single neuron stimulation!